move auth stuff to lib
