--- controllers/auth/utils.go
@@ -1,6 +1,0 @@
-package auth
-
-// TODO: make this hash well
-func hash(password string) string {
-	return password
-}
