+++ go.mod
@@ -3,7 +3,6 @@ module housemates/housemates-backend
 go 1.18
 
 require (
-	github.com/MicahParks/keyfunc v1.2.0
 	github.com/gin-gonic/gin v1.8.1
 	github.com/golang-jwt/jwt/v4 v4.4.2
 	github.com/joho/godotenv v1.4.0
